ml-services/python-ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
statsmodels==0.14.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
scipy==1.11.2